Added a few projects
Enumeration Script is a small script used to set a baseline for the critical parts of a Windows 10 computer.
index.html is a small Javascript project that uses RESTful calls to search TheMovieDB for movies that correlate to the user's input.
Logged.py is a Python project that was used to build spreadsheets based on user data from the INSPIRE game. All data was excluded from the script for user privacy and confidentiality.